Support Number group Items and number formatting
openHAB items like
Group:Number:AVG gHumidity    "Average Humidity [%.2f %%]" <humidity>
Group:Number:AVG gTemperature "Average Temperature [%.1f °C]" <temperature>

are now supported.

Also the formatting is supported now.

The [%.2f %%] will result in a number with 2 digits after the "." and a "%"
as the unit.

A [%.1f °C] will result in a number with 1 digit after the "." and a "°C"
as the unit.

A ["%d °C"] will result in an integer number and unit "°C".

Signed-off-by: Stefan Triller <github@stefantriller.de>